Several attempts to clarify instructions up to running the model.
I generated larger screen shots (and one for the updated main menu), as I
could not read the 300px wide images in the current site.  I am old, and
my eyesight is not what it was, but I think screen resolutions may have
got higher, too.

The main changes attempt to clarify the procedure for specifying the
covariates, which I had to work at.  I attempted to clarify by making
each covariate have its own paragraph and by repeating the instructions
for each covariate.  This may not be needed by the experienced, but I think
it may aid those who most need the tutorial material.

I am not able to get a full preview with the style sheets, so please adjust
image sizes as appropriate for correct display.

I hope this is useful.